move globals to a constants folder
