+++ src/main/java/com/karl/Main.java
@@ -1,6 +1,7 @@
 package com.karl;
 
 import java.io.IOException;
+import com.karl.constants.Globals;
 
 public class Main {
   public static void main(String[] args) throws IOException {
